Extract enricher mock and assertion validations
